fix: update interface for NFT
