+++ plugin-packages/model/src/gltf/index.ts
@@ -1,5 +1,4 @@
 export * from './loader-helper';
-export * from './loader-impl';
 export * from './protocol';
 export * from './json-converter';
-export * from './loader-ecs';
+export * from './loader-impl';
