+++ plugin-packages/model/src/gltf/protocol.ts
@@ -1,15 +1,9 @@
+import type { GLTFResources } from '@vvfx/resource-detection';
 import type {
-  GLTFMaterial, GLTFPrimitive, GLTFLight, GLTFScene, GLTFImage, GLTFTexture,
-  GLTFCamera, GLTFAnimation, GLTFImageBasedLight, GLTFResources,
-} from '@vvfx/resource-detection';
-import type { CubeImage } from '@vvfx/resource-detection/dist/src/gltf-tools/gltf-image-based-light';
-import type {
-  spec, Renderer, Texture, Geometry, TextureSourceOptions, EffectComponentData,
+  spec, TextureSourceOptions, EffectComponentData,
 } from '@galacean/effects';
 import type {
-  ModelAnimationOptions, ModelMaterialOptions, ModelSkyboxOptions, ModelTreeOptions,
-  ModelBaseItem, ModelLightComponentData, ModelCameraComponentData, ModelMeshComponentData,
-  ModelSkyboxComponentData,
+  ModelLightComponentData, ModelCameraComponentData, ModelSkyboxComponentData,
 } from '../index';
 
 /**
@@ -48,10 +42,6 @@ export interface LoadSceneOptions {
   },
   effects: {
     /**
-     * Renderer
-     */
-    renderer?: Renderer,
-    /**
      * 播放时间，单位秒
      */
     duration?: number,
@@ -72,15 +62,6 @@ export interface LoadSceneOptions {
 
 export interface LoadSceneResult {
   source: string,
-  items: ModelBaseItem[],
-  sceneAABB: {
-    min: spec.vec3,
-    max: spec.vec3,
-  },
-}
-
-export interface LoadSceneECSResult {
-  source: string,
   jsonScene: spec.JSONScene,
   sceneAABB: {
     min: spec.vec3,
@@ -88,50 +69,55 @@ export interface LoadSceneECSResult {
   },
 }
 
-export interface LoaderOptions {
-  compatibleMode?: 'gltf' | 'tiny3d',
-}
-
 export type SkyboxType = 'NFT' | 'FARM';
 
-export interface Loader {
-
-  loadScene (options: LoadSceneOptions): Promise<LoadSceneResult>,
-
-  processLight (lights: GLTFLight[], fromGLTF: boolean): void,
-
-  processCamera (cameras: GLTFCamera[], fromGLTF: boolean): void,
-
-  processMaterial (materials: GLTFMaterial[], fromGLTF: boolean): void,
-
-  createTreeOptions (scene: GLTFScene): ModelTreeOptions,
-
-  createAnimations (animations: GLTFAnimation[]): ModelAnimationOptions[],
-
-  createGeometry (primitive: GLTFPrimitive, hasSkinAnim: boolean): Geometry,
-
-  // 由于要兼容tiny开启了纹理预乘的模式，需要在外面创建和设置纹理，这里只做渲染Options对象的数据转换
-  createMaterial (material: GLTFMaterial): ModelMaterialOptions,
-
-  createTexture2D (images: GLTFImage, textures: GLTFTexture, isBaseColor: boolean): Promise<Texture>,
-
-  createTextureCube (cubeImages: CubeImage[], level0Size?: number): Promise<Texture>,
-
-  createSkybox (ibl: GLTFImageBasedLight): Promise<ModelSkyboxOptions>,
-
-  createDefaultSkybox (typeName: SkyboxType): Promise<ModelSkyboxOptions>,
-
-  // 用来转换 GLTF 和 Effects 之间颜色值范围
-  scaleColorVal (val: number, fromGLTF: boolean): number,
-
-  scaleColorVec (vec: number[], fromGLTF: boolean): number[],
-
+export interface ModelCamera {
+  fov: number,
+  near: number,
+  far: number,
+  clipMode: spec.CameraClipMode,
+  //
+  name: string,
+  position: spec.vec3,
+  rotation: spec.vec3,
+  duration: number,
+  endBehavior: spec.ItemEndBehavior,
 }
 
-export interface LoaderECS {
-  loadScene (options: LoadSceneOptions): Promise<LoadSceneECSResult>,
+export interface ModelLight {
+  lightType: spec.LightType,
+  color: spec.ColorData,
+  intensity: number,
+  range?: number,
+  innerConeAngle?: number,
+  outerConeAngle?: number,
+  //
+  name: string,
+  position: spec.vec3,
+  rotation: spec.vec3,
+  scale: spec.vec3,
+  duration: number,
+  endBehavior: spec.ItemEndBehavior,
+}
 
-  processGLTFResource (resource: GLTFResources): void,
+export interface ModelSkybox {
+  skyboxType?: string,
+  renderable?: boolean,
+  intensity?: number,
+  reflectionsIntensity?: number,
+  duration?: number,
+}
+
+export interface ModelImageLike {
+  name?: string,
+  width: number,
+  height: number,
+}
+
+export interface Loader {
+  loadScene (options: LoadSceneOptions): Promise<LoadSceneResult>,
+
+  processGLTFResource (resource: GLTFResources, imageElements: ModelImageLike[]): void,
 
   processComponentData (components: EffectComponentData[]): void,
 
@@ -139,11 +125,9 @@ export interface LoaderECS {
 
   processCameraComponentData (camera: ModelCameraComponentData): void,
 
-  processMeshComponentData (mesh: ModelMeshComponentData): void,
-
   processSkyboxComponentData (skybox: ModelSkyboxComponentData): void,
 
   processMaterialData (material: spec.MaterialData): void,
 
-  processTextureOptions (options: TextureSourceOptions, isBaseColor: boolean): void,
+  processTextureOptions (options: TextureSourceOptions, isBaseColor: boolean, image?: ModelImageLike): void,
 }
