+++ plugin-packages/model/src/helper/index.ts
@@ -1,3 +1,2 @@
 export * from '../gltf/loader-ext';
-export * from '../gltf/loader-ecs-ext';
 export * from '../gesture/controller';
